Cleaned up and Refactored
Did some slight refactoring and have the code pretty much fully cleaned up. Also with the refactoring brought down the training time from 90 seconds, to 80, and now getting around 38 seconds consistently, with the same loss, its much faster! But that's only for 10 epochs, still a long a way to go