RECURSION ADDED LFGOOOOO
After many iterations and some breaks for optimization testing, I decided that any complicated optimization methods would distract from the simplicity, readability, and comprehension of the program, so I bit the bullet and finalized the recursion method, perfectly replicating the hardcoded method I was using before and receiving the same outputs as before. I also got the standardized "Left to right" reading for the activation function scripts. It even still works for having no hidden layers at all, as well as having support for having no activation function in the hidden layers, if one so chooses. Very excited to get this made and think we should start working on getting this ready for a package release filled with educational material ^^. I will of course keep moving forward with optimization, but I feel the state it's in is best for understanding how everything works with an honestly quite small amount of code. I have some ideas for OpenGL integration and did a little messing around in pure C. But again, that will wait for this first edition to be completed.